Removed unnecessary requests to database
