+++ classes/Post.php
@@ -106,6 +106,20 @@ class Post
         }
     }
 
+    public function getPosts() {
+        if ($stmt = $this->mysqli->query("SELECT post_id, account_id, title, code, subtext, timestamp FROM posts ORDER BY timestamp DESC")) {
+            if ($stmt->num_rows > 0) {
+                while ($row = $stmt->fetch_array(MYSQLI_ASSOC)) {
+                    $test[] = $row;
+                }
+                return $test;
+            }
+        } else {
+            return $this->mysqli->error;
+        }
+    }
+
+
     private function updatePoints($username, $points) {
         $stmt = $this->mysqli->prepare("UPDATE account SET punten = punten + ? WHERE username = ?");
         $stmt->bind_param("is", $points, $username);
